Entity archetype swap
There was a issue where when adding components and the the entity which we where adding components to moved to its new archetype, it wasnt at the final row meaning. The source archetype had to perform an internal swap resulting in non pod data types getting mixed. fixed for now